Share the same property dictionary between LoggerActivity.Captures (now Properties) and ActivityInstrumentation.
