+++ src/SerilogTracing.Sinks.OpenTelemetry/SerilogTracing.Sinks.OpenTelemetry.csproj
@@ -1,16 +1,14 @@
 <Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<Description>This Serilog sink transforms Serilog events into OpenTelemetry
-			logs and sends them to an OTLP (gRPC or HTTP) endpoint.</Description>
-		<VersionPrefix>2.0.0</VersionPrefix>
+		<Description>Sends log events and traces to OTLP (gRPC or HTTP) endpoints.</Description>
 		<Authors>Serilog Contributors</Authors>
 		<LangVersion>10</LangVersion>
 		<TargetFramework>net6.0</TargetFramework>
 		<PackageTags>serilog;sink;opentelemetry</PackageTags>
 		<PackageIcon>serilog-sink-nuget.png</PackageIcon>
-		<PackageProjectUrl>https://github.com/serilog/serilog-sinks-opentelemetry</PackageProjectUrl>
+		<PackageProjectUrl>https://github.com/serilog-tracing/serilog-tracing</PackageProjectUrl>
 		<PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
-		<RepositoryUrl>https://github.com/serilog/serilog-sinks-opentelemetry</RepositoryUrl>
+		<RepositoryUrl>https://github.com/serilog-tracing/serilog-tracing</RepositoryUrl>
 		<RepositoryType>git</RepositoryType>
 		<GenerateDocumentationFile>true</GenerateDocumentationFile>
 		<RootNamespace>Serilog</RootNamespace>
