+++ src/SerilogTracing/SerilogTracing.csproj
@@ -6,7 +6,9 @@
     <PackageReadmeFile>README.md</PackageReadmeFile>
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     <PackageTags>serilog;tracing</PackageTags>
-    <Description>An experimental Serilog extension for producing and capturing hierarchical traces.</Description>
+    <Description>SerilogTracing is a minimal tracing system that integrates Serilog with System.Diagnostics.Activity.
+      You can use it to add distributed, hierarchical tracing to applications that use Serilog, and to consume traces
+      generated by .NET components including HttpClient and ASP.NET Core.</Description>
   </PropertyGroup>
   
   <ItemGroup>
